+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedAppsDashboard/usePurchasedOrders.ts
@@ -25,28 +25,33 @@ const usePurchasedOrders = ({
 	const key = `/placed-orders/${accountId}/${channelId}/${page}/${pageSize}`;
 
 	const swr = useSWR(
-		`/placed-orders/${accountId}/${channelId}/${page}/${pageSize}`,
+		accountId && channelId
+			? `/placed-orders/${accountId}/${channelId}/${page}/${pageSize}`
+			: null,
 		async () => {
-			const placedOrders = await HeadlessCommerceDeliveryOrder.getPlacedOrders(
-				channelId,
-				accountId,
-				new URLSearchParams({
-					nestedFields: 'placedOrderItems',
-					page: page.toString(),
-					pageSize: pageSize.toString(),
-				})
+			const placedOrders =
+				await HeadlessCommerceDeliveryOrder.getPlacedOrders(
+					channelId,
+					accountId,
+					new URLSearchParams({
+						nestedFields: 'placedOrderItems',
+						page: page.toString(),
+						pageSize: pageSize.toString(),
+					})
+				);
+
+			const placedOrdersFiltered = placedOrders.items.filter(
+				({orderTypeExternalReferenceCode}) =>
+					orderTypeExternalReferenceCodes.length
+						? orderTypeExternalReferenceCodes.includes(
+								orderTypeExternalReferenceCode
+						  )
+						: true
 			);
 
 			return {
 				...placedOrders,
-				items: placedOrders.items.filter(
-					({orderTypeExternalReferenceCode}) =>
-						orderTypeExternalReferenceCodes.length
-							? orderTypeExternalReferenceCodes.includes(
-									orderTypeExternalReferenceCode
-							  )
-							: true
-				),
+				items: placedOrdersFiltered,
 			};
 		}
 	);
